+++ test/general-spec.js
@@ -12,11 +12,13 @@ eval('var syncFunction = ' + fs.readFileSync('build/sync-functions/test-general-
 // Placeholders for stubbing built-in Sync Gateway support functions.
 // More info: http://developer.couchbase.com/mobile/develop/guides/sync-gateway/sync-function-api-guide/index.html
 var requireAccess;
+var requireRole;
 var channel;
 
 describe('Functionality that is common to all documents:', function() {
   beforeEach(function() {
     requireAccess = simple.stub();
+    requireRole = simple.stub();
     channel = simple.stub();
   });